separate the component into functional and rendering parts
